Fixing boolean parsing
After fixing the lexer to recognize TRUE/FALSE this exposed a bug in
the parser where we weren't setting the value correctly. The parser
was trying to call a method called `this.curToken(token.TRUE)`, this
is not a method but a property of the parser.

After fixing this there were some unit tests that were poorly written
that had to be adjusted